Include CloudWatch event limits
According to their documentation, "the maximum batch size is 1,048,576
bytes. This size is calculated as the sum of all event messages in
UTF-8, plus 26 bytes for each log event." Currently we only include the
first part of size calculation. This means that any batch with more than
1868 messages will fail with an invalid parameter error because we only
allow 48,576 extra bytes. Now, we'll allow you to send batches with any
number of events as long as they pass AWS's limits.