Stop memory leaking CloudWatch sequence tokens in cloudwatch-integration
The _nextToken variable saved every sequence token for every stream seen.
Since cloudwatch-integration stores _nextToken as a top level variable,
it gets shared between all instances of WinstonCloudWatch and leaks
memory.

To fix this, add a clearSequenceToken method and call it from kthxbye to
delete the entry for the WinstonCloudWatch instance that is done sending
logs.